+++ packages/react-native-nitro-markdown/android/src/main/cpp/cpp-adapter.cpp
@@ -1,7 +1,10 @@
 #include <jni.h>
+#include <fbjni/fbjni.h>
 #include "NitroMarkdownOnLoad.hpp"
 
 JNIEXPORT jint JNICALL JNI_OnLoad(JavaVM* vm, void*) {
-  return margelo::nitro::Markdown::initialize(vm);
+  return facebook::jni::initialize(vm, []() {
+    margelo::nitro::Markdown::registerAllNatives();
+  });
 }