+++ packages/react-native-nitro-markdown/android/src/main/java/com/nitromarkdown/NitroMarkdownPackage.kt
@@ -1,13 +1,13 @@
 package com.nitromarkdown
 
-import com.facebook.react.TurboReactPackage
+import com.facebook.react.BaseReactPackage
 import com.facebook.react.bridge.NativeModule
 import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.module.model.ReactModuleInfoProvider
 import com.facebook.react.uimanager.ViewManager
 import com.margelo.nitro.com.nitromarkdown.NitroMarkdownOnLoad
 
-class NitroMarkdownPackage : TurboReactPackage() {
+class NitroMarkdownPackage : BaseReactPackage() {
     companion object {
         init {
             NitroMarkdownOnLoad.initializeNative()